Move Volume and GlobalVolume to own file
